+++ bzt/engine.py
@@ -916,7 +916,7 @@ class ScenarioExecutor(EngineModule):
         :return: DictOfDicts
         """
         if self.__scenario is None:
-            scenario = self.execution.get('scenario', {})
+            scenario = self.execution.get('scenario', BetterDict())
             if isinstance(scenario, string_types):
                 scenarios = self.engine.config.get("scenarios")
                 if scenario not in scenarios: