+++ tests/locust/simple.py
@@ -1,0 +1,23 @@
+from locust import HttpLocust, TaskSet, task
+
+
+class WebsiteTasks(TaskSet):
+    def on_start(self):
+        self.client.post("/login", {
+            "username": "test_user",
+            "password": ""
+        })
+
+    @task
+    def index(self):
+        self.client.get("/")
+
+    @task
+    def about(self):
+        self.client.get("/about/")
+
+
+class WebsiteUser(HttpLocust):
+    task_set = WebsiteTasks
+    min_wait = 100
+    max_wait = 1500